# Refactor set.SliceSubsetOf