--- a/pkg/set/doc.go
+++ b/pkg/set/doc.go
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Authors of Cilium
-
-// Package set contains a function for performing a subset check for slices.
-package set