pkg/set has a method to verify if a given slice is a subset of another. Being an utility function about slices, it is moved in the pkg/slices package.
Also, the method is modified to leverage type parameters, and broken down to expose the diff logic as a separate method, to higher the chances of being reused throughout the codebase.

The algorithm used for the new `slices.Diff` and `slices.SubsetOf` differ from the previous semantic in a detail: they do not consider duplicates. IOW, the input slices are treated as sets. This is fine with the current usage of the method and in case the opposite semantic will be needed in the future, it'll be possible to add a new method with a different algorithm.