@pippolo84 2023-05-24
> Nice! Benchmarks between the new implementation and the one it replaces?

I've added a commit with a benchmark.
Here the results (I've also added this in the commit message):

```
    name                  old time/op    new time/op    delta
    SubsetOf/64-512-8       15.3µs ±35%    14.6µs ±27%      ~     (p=0.134 n=20+20)
    SubsetOf/128-512-8      24.4µs ± 5%    22.6µs ± 4%    -7.47%  (p=0.000 n=18+17)
    SubsetOf/256-2048-8     88.3µs ± 7%    81.0µs ± 6%    -8.28%  (p=0.000 n=18+19)
    SubsetOf/512-2048-8      110µs ± 5%      98µs ± 7%   -10.24%  (p=0.000 n=20+20)
    SubsetOf/1024-8192-8     358µs ± 4%     318µs ± 5%   -11.25%  (p=0.000 n=19+20)
    SubsetOf/2048-8192-8     469µs ± 5%     423µs ±10%    -9.72%  (p=0.000 n=20+19)
    
    name                  old alloc/op   new alloc/op   delta
    SubsetOf/64-512-8       28.7kB ± 0%    23.2kB ± 0%   -19.10%  (p=0.000 n=20+20)
    SubsetOf/128-512-8      28.8kB ± 0%    25.9kB ± 0%    -9.97%  (p=0.000 n=19+16)
    SubsetOf/256-2048-8      115kB ± 0%      92kB ± 0%   -19.62%  (p=0.000 n=19+20)
    SubsetOf/512-2048-8      115kB ± 0%     103kB ± 0%   -10.78%  (p=0.000 n=19+16)
    SubsetOf/1024-8192-8     459kB ± 0%     360kB ± 0%   -21.42%  (p=0.000 n=18+17)
    SubsetOf/2048-8192-8     460kB ± 0%     402kB ± 0%   -12.59%  (p=0.000 n=20+17)
    
    name                  old allocs/op  new allocs/op  delta
    SubsetOf/64-512-8         2.00 ± 0%      4.00 ± 0%  +100.00%  (p=0.000 n=20+20)
    SubsetOf/128-512-8        4.65 ±57%      5.55 ±28%   +19.35%  (p=0.028 n=20+20)
    SubsetOf/256-2048-8       2.00 ± 0%      4.00 ± 0%  +100.00%  (p=0.000 n=19+20)
    SubsetOf/512-2048-8       6.45 ±24%      7.75 ±10%   +20.16%  (p=0.000 n=20+16)
    SubsetOf/1024-8192-8      2.17 ±38%      4.00 ± 0%   +84.62%  (p=0.000 n=18+17)
    SubsetOf/2048-8192-8      8.65 ± 8%     10.00 ± 0%   +15.61%  (p=0.000 n=20+17)
```

The number of allocations increased, but the total amount of allocated memory lowered.
Also, there is an improvement in speed, due to the simpler semantic.